+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - removed `remoteConfig().setDefaults({})` & replaced with `remoteConfig().defaultConfig = {}` which is a web property.
 - removed `remoteConfig().setDefaultsFromResource({})`, not in the web spec.
 - removed `remoteConfig().setConfig({})` & replaced with `remoteConfig().settings = {}` which is a web property.
+- removed `settings.minimumFetchInterval` in JS & changed it to `settings.minimumFetchIntervalMillis` as per web spec.
+- added `settings.fetchTimeMillis` as per web spec.
 
 # [6.1.0](https://github.com/invertase/react-native-firebase/compare/v6.0.4...v6.1.0) (2019-11-26)